Adds stripe checkout functionality
Moves all CC information into a form provided by Stripe. This form has
lots of built-in fanciness and is better than the homegrown one we had
previously. On successful payment, and POST request is sent to our API
which includes the token from the interaction with Stripe. The customer
is not charged until the order is successfully created on the server.