fixes issue identified by instructional staff
Issue description from Mike F.:
User facing bugs
- If you add an item twice, there is no user feedback

After investigating the bug, it appears that users are not receiving
feedback when trying to add to the cart when they have signed in then
signed back out of the site. They are receiving feedback when trying to
add an item twice when they are signed in and before they have signed in
after first visitng the site.

This commit fixes the bug by updating the if statement in carts/events.js
addToCart function.