Updates sign in check when trying to open cart
Adds check for empty object when trying to access cart. Previously, users
would not receive feedback when trying to access cart after signing out.
This is because the user.store object was being checked to see if user is
logged in. User.store exists after logout, but is empty. Empty object
check helps with this.