Updates product clickhandlers
Changes the way clickhandlers are added to products which are displayed
using handlebars templates. Product ID is now set as the id element
for the add product buttons. This ID is used to know what product to add
to the cart. This fixes an issue identified by the instructional staff.